requirements.txt
streamlit
pandas
plotly
openpyxl
scikit-learn
gdown 
pyarrow 
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client